Moved .NET 5 tests for BunitJSInterop to their own file and made that test class partial
